+++ error_propagation/src/finance/npv.py
@@ -4,7 +4,7 @@ from typing import Union
 
 import numpy as np
 
-from error_propagation.complex import Complex
+from error_propagation.src.complex import Complex
 
 
 def npv(
