+++ setup.py
@@ -8,5 +8,6 @@ setup(
     author_email="davidreissmello@gmail.com",
     description="Package that computes error propagation",
     url="https://github.com/davidreissmello/error_propagation",
-    packages=find_packages(where=""),
+    package_dir={"": "src"},
+    packages=find_packages(where="src"),
 )
