+++ tests/test_complex.py
@@ -3,7 +3,7 @@ from math import sqrt
 
 import pytest
 
-from error_propagation.complex import Complex
+from error_propagation.src.complex import Complex
 
 
 class TestComplexClass:
@@ -13,13 +13,13 @@ class TestComplexClass:
     def test_complex_value_error_output(self):
         """Check that inputs other than int or float return a value error"""
         with pytest.raises(ValueError):
-            Complex(value='hello', error=3)
+            Complex(value="hello", error=3)
 
         with pytest.raises(ValueError):
-            Complex(value=3, error='world')
+            Complex(value=3, error="world")
 
         with pytest.raises(ValueError):
-            Complex(value='hello', error='world')
+            Complex(value="hello", error="world")
 
 
 class TestOperations:
@@ -41,14 +41,11 @@ class TestOperations:
 
     def test_multiplication(self):
         result = Complex(6, 7) * Complex(8, 9)
-        assert result == Complex(
-            6 * 8,
-            sqrt(6 ** 2 * 9 ** 2 + 8 ** 2 * 7 ** 2),
-        )
+        assert result == Complex(6 * 8, sqrt(6 ** 2 * 9 ** 2 + 8 ** 2 * 7 ** 2))
 
     def test_division(self):
         result = Complex(8, 7) / Complex(4, 3)
-        expected_error = sqrt((8 ** 2 * 3 ** 2)/(4 ** 4) + (7 ** 2)/(4 ** 2))
+        expected_error = sqrt((8 ** 2 * 3 ** 2) / (4 ** 4) + (7 ** 2) / (4 ** 2))
         assert result == Complex(2, expected_error)
 
     def test_power(self):
@@ -59,8 +56,8 @@ class TestOperations:
 
         result = Complex(a, a_error) ** Complex(b, b_error)
         expected_error = sqrt(
-            a ** (2 * b) * b_error ** 2 * log(a) ** 2 +
-            (a ** (2 * b) * a_error ** 2 * b ** 2)/(a ** 2),
+            a ** (2 * b) * b_error ** 2 * log(a) ** 2
+            + (a ** (2 * b) * a_error ** 2 * b ** 2) / (a ** 2)
         )
 
         assert result == Complex(a ** b, expected_error)
